Two features ready for review
I have created the functionality for both importing an image into blender and saving the current file in blender. Plugin.py has all of the functionality in one file while the save.py just has the save functionality. Something to note is that the importing images will only work with the front image and only if the front image has a file destination ready. More steps will be needed to import different or all images although the code is all there you would just have to copy and paste for the different views.