Updated File conversion
I have updated file conversion to include the decoding section for the database data. It currently doesn't work because the data that is returned from decoding is garbage. On top of this there will most likely be another parameter that will be added to change the type of temp file.