Updated file_conversion
Updated file_conversion to include decoding data from the database. This functionality does not work, but that is because the decoded data is garbage. The temp file is created, but does not load. There will also have to be another parameter to the method to change the temp file type.