Updated file_conversion
Updated the file_conversion to include the .blend flies. Currently does not work with Windows11 naming scheme for the users folders, as they hyphenate the name instead of adding in the spaces. The next addition to this file will be concatenating the strings so that we can take care of this.