Added the base Depth By Colors Methodology
This method should go through the points from in the image and find the outside points of the shape.
We then get the line data for all of these shapes so we can calculate how much around the point we should account for. Finally we take the average of the colors surrounding the point and normalize those values to get a list of Z axis points.